Initial Cleanup, and Project Setup
