+++ platforms/android/app/build.gradle
@@ -348,3 +348,7 @@ if (hasProperty('postBuildExtras')) {
 if (cordovaConfig.IS_GRADLE_PLUGIN_GOOGLE_SERVICES_ENABLED) {
     apply plugin: 'com.google.gms.google-services'
 }
+
+task prepareKotlinBuildScriptModel {
+
+}
